sign up set up nowworking with netlify and gotrue
