+++ src/Components/App/App.jsx
@@ -1,5 +1,5 @@
 import AllRecipes from "../../views/AllRecipes/AllRecipes";
-import SignUp from "../LoginSetUpLoreen/LoginSetUp";
+import SignUp from "../LoginSetUpLoreen/SignSetUp";
 const App = () => {
   return (
     <>
